fix generic page
