+++ src/App.js
@@ -1,5 +1,5 @@
 import React, { useState } from 'react';
-import Home from './pages/Home';
+import Home from './pages/generic page/Home';
 import Detail from './pages/Detail';
 import Listing from './pages/Listing';
 import MailInfo from './pages/MailInfo';
