--- src/pages/generic page/Generic.scss
@@ -1,36 +1,0 @@
-.Generic {
-
-  display: flex;
-  flex-direction: column;
-  width: 70%;
-  margin: auto;
-
-  h1 {
-    
-  }
-
-  .section {
-    display: flex;
-    flex-direction: column;
-    justify-content: flex-start;
-    background-color: white;
-    border: 1px solid lightgray;
-    box-shadow: lightgray;
-    border-radius: 5px;
-    margin-bottom: 25px;
-    text-align: left;
-
-    h2 {
-      padding-left: 10px;
-      margin: 15px 0 10px 0;
-    }
-
-    p {
-      padding-left: 10px;
-      margin-top: 5px;
-      margin-bottom: 15px;
-    }
-
-  }
-
-}
