+++ src/pages/generic page/Home.scss
